CAR PORT: 2019 Lexus LS 500
This is my first attempt to port a car.

Values in `interface.py` were copied and modified from `LEXUS_ES_TSS2`. `wheelbase` was calculated from the press release value for the base LS 500 (123 inches). The `tire_stiffness_factor` value likely needs to be increased due to the stock tires being run flats. `mass` is from the press release value for the base LS 500 with air suspension.

https://pressroom.lexus.com/vehicle/2019-lexus-ls/